update windows build

+++ build_windows.py
@@ -19,7 +19,8 @@ subprocess.run(["flutter", "build", "windows"], shell=True)
 file = open('pubspec.yaml', 'w')
 file.write(content)
 
-os.remove("build/app-windows.zip")
+if os.path.exists("build/app-windows.zip"):
+    os.remove("build/app-windows.zip")
 
 # 压缩build/windows/x64/runner/Release, 生成app-windows.zip, 使用tar命令
 subprocess.run(["tar", "-a", "-c", "-f", "build/app-windows.zip", "-C", "build/windows/x64/runner/Release", "."]
